Moved plugin registry back to pkg/epp/plugins
